// src/app/models/data-store.ts
export enum SessionDataStore {
    email = 'email',
    phone = 'phone',
    productId = "productId",
    brandId = 'brandId',
    productName = "product",
    brandName = 'brand',
    supportAmount = 'supportAmount',
    organizationId = 'organizationId',
    selectedProductRepairCodes = 'selectedProductRepairCodes',
    selectedProductSymptoms = 'selectedProductSymptoms',
    claimId = 'claim-id',
    invoice = 'invoice',
    errors = "errors",
    key = 'key',
    validatores = 'validatores',
}